Rename Cockroach to CockroachDB
Rename Cockroach to CockroachDB
