+++ src/components/techSwitcher.tsx
@@ -9,7 +9,7 @@ import MongoDB from '../icons/technologies/MongoDB'
 import JS from '../icons/technologies/JS'
 import SQLServer from '../icons/technologies/MSSQL'
 import PlanetScale from '../icons/technologies/PlanetScale'
-import Cockroach from '../icons/technologies/Cockroach'
+import CockroachDB from '../icons/technologies/CockroachDB'
 import { components } from 'react-select'
 import ArrowDown from '../icons/ArrowDown'
 
@@ -33,12 +33,12 @@ const icons = {
   mongodb: <MongoDB />,
   sqlserver: <SQLServer />,
   planetscale: <PlanetScale />,
-  cockroach: <Cockroach />,
+  cockroachdb: <CockroachDB />,
 }
 
 const technologyTypes = {
   languages: ['node', 'typescript'],
-  databases: ['mysql', 'postgres', 'sqlite', 'mongodb', 'sqlserver', 'planetscale', 'cockroach'],
+  databases: ['mysql', 'postgres', 'sqlite', 'mongodb', 'sqlserver', 'planetscale', 'cockroachdb'],
 }
 
 export const technologyNames = {
@@ -50,7 +50,7 @@ export const technologyNames = {
   mongodb: 'MongoDB',
   sqlserver: 'SQL Server',
   planetscale: 'PlanetScale',
-  cockroach: 'Cockroach',
+  cockroachdb: 'CockroachDB',
 }
 
 const TechnologySwitch = ({ type, onChangeTech, technologies, defaultTech }: TechSwitchProps) => {
