+++ src/icons/technologies/CockroachDB.tsx
